# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Be
  description: |
    The API provides CSR backend API.
basePath: /api/
schemes:
  - https
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /v1/orders/status/{status}:
    get:
      summary: Get orders by status.
      tags:
        - Orders
      security:
        - Bearer: [ ]
      parameters:
        - name: status
          in: path
          required: true
          description: order status
          type: string
        - name: limit
          in: query
          required: false
          description: limit of items in page
          type: integer
          default: 10
        - name: offset
          in: query
          required: false
          description: offset of items
          type: integer
          default: 0
        - name: order_by
          in: query
          required: false
          description: how to order list
          type: string
          enum:
            - asc
            - desc
          default: asc
        - name: order_column
          in: query
          required: false
          description: column to order by
          type: string
          enum:
            - id
            - created_at
            - rent_start
            - rent_end
            - is_first
          default: id
      operationId: GetOrdersByStatus
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/UserOrdersList"
        404:
          description: No orders found
          schema:
            type: string
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/SwaggerError'
...
  /v1/status_names:
    get:
      summary: Get all status names.
      tags:
        - Orders
      security:
        - Bearer: [ ]
      operationId: GetAllStatusNames
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/OrderStatusNames"
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/SwaggerError'
...
definitions:
...
  SwaggerError: # Do not delete it. It is used for generating error responses.
    type: object
    required:
      - code
      - message
    properties:
        code:
            type: integer
            format: int32
        message:
            type: string
        details:
            type: string
  PetKind:
    type: object
    required:
      - name
    properties:
      name:
        type: string
...
  UserEmbeddable:
    type: object
    required:
      - id
      - name
    properties:
      id:
        type: integer
      name:
        type: string
...
  EquipmentResponse:
    type: object
    required:
      - id
      - termsOfUse
      - name
      - title
      - category
      - status
      - petSize
      - compensationCost
      - technicalIssues
      - inventoryNumber
      - supplier
      - receiptDate
      - maximumDays
      - location
      - description
      - photoID
    properties:
      id:
        type: integer
        example: 1
      termsOfUse:
        type: string
        example: "https://..."
      name:
        type: string
        example: "Dog harness 3000"
      title:
        type: string
        example: "клетка midwest icrate 1"
      category:
        type: integer
        example: 1
      subcategory:
        type: integer
        example: 1
      status:
        type: integer
        example: 1
      petSize:
        type: integer
        example: 1
      compensationCost:
        type: integer
        example: 3900
      technicalIssues:
        type: boolean
        example: false
        x-nullable: true
      condition:
        type: string
        example: "удовлетворительное, местами облупляется краска"
      inventoryNumber:
        type: integer
        example: "1"
      supplier:
        type: string
        example: "ИП Григорьев Виталий Васильевич"
      receiptDate:
        type: integer
        example: 1678111533
      maximumDays:
        type: integer
        example: 30
      location:
        type: integer
        example: 71
      description:
        type: string
        example: "This is a dog harness.\nWARNING: do not put on cats!"
      photoID:
        type: string
      petKinds:
        type: array
        items:
          $ref: "#/definitions/PetKind"
...
  UserOrder:
    type: object
    required:
      - id
      - description
      - quantity
      - rent_start
      - rent_end
      - user
      - equipments
      - last_status
      - is_first
    properties:
      id:
        type: integer
      description:
        type: string
      quantity:
        type: integer
      rent_start:
        type: string
        format: date-time
      rent_end:
        type: string
        format: date-time
      user:
        type: object
        $ref: '#/definitions/UserEmbeddable'
      equipments:
        type: array
        items:
          $ref: '#/definitions/EquipmentResponse'
      last_status:
        type: object
        $ref: '#/definitions/OrderStatus'
      is_first:
        type: boolean
        example: false
...
  OrderStatus:
    type: object
    required:
      - id
      - comment
      - status
      - created_at
      - changed_by
      - order_id
    properties:
      id:
        type: integer
      comment:
        type: string
      status:
        type: string
      created_at:
        type: string
        format: date-time
      changed_by:
        type: object
        $ref: '#/definitions/UserEmbeddable'
      order_id:
        type: integer
  UserOrdersList:
    type: object
    required:
      - total
      - items
    properties:
      total:
        type: integer
      items:
        type: array
        items:
          $ref: "#/definitions/UserOrder"
...
  OrderStatusNames:
    type: array
    items:
      $ref: '#/definitions/OrderStatusName'
  OrderStatusName:
    type: object
    required:
      - id
      - name
    properties:
      id:
        type: integer
      name:
        type: string